#43 Move Term outside CQL3
